UPDATED routing of error report endpoints.
