+++ src/app/find/Router.jsx
@@ -4,6 +4,7 @@ import { Route } from 'react-router-dom';
 import MapPageContainer from './mapPage/MapPageContainer';
 import QuestionFlowContainer from './questionFlow/QuestionFlowContainer';
 import LocationDetailsContainer from './locationDetails/LocationDetailsContainer';
+import ErrorReportContainer from './errorReports/ErrorReportContainer';
 
 export default function Router({ match }) {
   return (
@@ -15,6 +16,10 @@ export default function Router({ match }) {
       />
 
       <Route path={`${match.path}/location/:locationId`} component={LocationDetailsContainer} />
+      <Route
+        path={`${match.path}/location/:locationId/errorreports`}
+        component={ErrorReportContainer}
+      />
       <Route
         path={`${match.path}/:categoryName/location/:locationId`}
         component={LocationDetailsContainer}
